change mip-3d-gltf.vue to js
