+++ components/mip-3d-gltf/example/mip-3d-gltf.html
@@ -18,7 +18,7 @@
 
     <h2>1、通过src远程获取模型</h2>
     <mip-3d-gltf
-      layout="responsive"
+      layout="fixed"
       width="250"
       height="450"
       src="https://ampbyexample.com/glTF/DamagedHelmet.glb">
@@ -30,7 +30,7 @@
       width="250"
       height="450"
       alpha="true"
-      maxPixelRatio="0.5"
+      max-pixel-ratio="3"
       clear-color="#FFFFFF"
       antialiasing="true"
       enable-zoom="false"
